+++ CHANGELOG.md
@@ -13,3 +13,13 @@ Any entries below v1.0.0 will not be useful 😅😅
 ## [0.0.3] - 14/8/2019
 
 * Some tweaking.
+
+## [0.0.4-dev.1] - 29/8/2019
+
+* Rewrite to support plain dart without flutter
+
+## [1.0.0] - 29/8/2019
+
+* Stable release
+* Shipped Lipa Na Mpesa method
+* Improved docs